+++ selectrum.el
@@ -239,15 +239,15 @@ with the string the user inserted."
 
 Possible values are:
 
-- \\='matches: Show the total number of matches.
-- \\='current/matches: Show the index of current match and the total number of
-  matches.
+- `matches': Show the total number of matches.
+- `current/matches': Show the index of current match and the
+  total number of matches.
 - nil: Show nothing."
   :type '(choice
           (const :tag "Disabled" nil)
-          (const :tag "Count matches" 'matches)
+          (const :tag "Count matches" matches)
           (const :tag "Count matches and show current match"
-                 'current/matches)))
+                 current/matches)))
 
 (defcustom selectrum-show-indices nil
   "Non-nil means to number the candidates (starting from 1).