+++ CHANGELOG.md
@@ -135,6 +135,8 @@ The format is based on [Keep a Changelog].
   minibuffer instead of signaling an error and erasing the minibuffer
   contents ([#193]).  If ‘completion-fail-discreetly’ is non-nil,
   nothing is done.
+* Fix type mismatch when configuring `selectrum-count-style` in
+  customizations.
 * Submitting the default value with the empty prompt does no longer
   strip the text properties of the default candidate ([#180], [#198]).